#14: used the same map background for phase 1 & 2
